finish task 2 - add 404 errors - fix params types

+++ src/api/recipes.ts
@@ -5,7 +5,7 @@ import { eq } from "drizzle-orm";
 
 const Id = z.int32().min(0);
 
-const RecipeRead = z.object({
+export const RecipeRead = z.object({
   id: Id,
   title: z.string().max(200),
   description: z.string(),
@@ -41,7 +41,19 @@ const RecipeCreate = z.object({
   }).array(),
 });
 
-const baseQuery = {
+const Params = z.object({
+  id: z.coerce.number(),
+});
+
+const Message = z.object({
+  message: z.string(),
+});
+
+const Error = z.object({
+  error: z.string(),
+});
+
+export const baseQuery = {
   with: {
     cuisine: true,
     recipeAllergens: {
@@ -57,7 +69,7 @@ const baseQuery = {
   },
 } satisfies Parameters<typeof db.query.recipes.findFirst>[0];
 
-function serializeRecipe(recipe: NonNullable<Awaited<ReturnType<typeof db.query.recipes.findFirst<typeof baseQuery>>>>): z.infer<typeof RecipeRead> {
+export function serializeRecipe(recipe: NonNullable<Awaited<ReturnType<typeof db.query.recipes.findFirst<typeof baseQuery>>>>): z.infer<typeof RecipeRead> {
   const { cuisineId, recipeAllergens, recipeIngredients, ...rest } = recipe;
 
   return {
@@ -113,28 +125,38 @@ export const recipesRouter = new Elysia({ prefix: "/recipes" })
     body: RecipeCreate,
     response: { 201: RecipeRead },
   })
-  .get("/:id", async ({ query: { id } }) => {
+  .get("/:id", async ({ params: { id }, status }) => {
     const recipe = await db.query.recipes.findFirst({
       ...baseQuery,
       where: eq(schema.recipes.id, id),
     });
-    return recipe && serializeRecipe(recipe);
+    if (!recipe) return status(404, {
+      error: "recipe with this id not found",
+    });
+    return serializeRecipe(recipe);
   }, {
-    query: z.object({ id: z.number() }),
-    response: RecipeRead.optional(),
+    params: Params,
+    response: {
+      200: RecipeRead,
+      404: Error,
+    },
   })
-  .put("/:id", async ({ query: { id }, body }) => {
+  .put("/:id", async ({ params: { id }, body, status }) => {
     return await db.transaction(async (tx) => {
-      await tx
+      const updated = await tx
         .update(schema.recipes)
         .set(body)
-        .where(eq(schema.recipes.id, id))
+        .where(eq(schema.recipes.id, id));
+
+      if (updated.rowsAffected === 0) return status(404, {
+        error: "recipe with this id not found",
+      });
 
       await tx.delete(schema.recipeAllergens).where(eq(schema.recipeAllergens.recipeId, id));
       await tx.insert(schema.recipeAllergens).values(
         body.allergenIds.map((allergenId) => ({
           recipeId: id,
-          allergenId
+          allergenId,
         }))
       );
 
@@ -156,11 +178,14 @@ export const recipesRouter = new Elysia({ prefix: "/recipes" })
       return serializeRecipe(recipe);
     });
   }, {
-    query: z.object({ id: z.number() }),
+    params: Params,
     body: RecipeCreate,
-    response: RecipeRead.optional(),
+    response: {
+      200: RecipeRead,
+      404: Error,
+    },
   })
-  .delete("/:id", async ({ query: { id }, status }) => {
+  .delete("/:id", async ({ params: { id }, status }) => {
     return await db.transaction(async (tx) => {
       await tx
         .delete(schema.recipeAllergens)
@@ -170,13 +195,22 @@ export const recipesRouter = new Elysia({ prefix: "/recipes" })
         .delete(schema.recipeIngredients)
         .where(eq(schema.recipeIngredients.recipeId, id));
 
-      await tx
+      const deleted = await tx
         .delete(schema.recipes)
         .where(eq(schema.recipes.id, id));
 
-      return status(204, {});
+      if (deleted.rowsAffected === 0) return status(404, {
+        error: "recipe with this id not found",
+      });
+
+      return status(204, {
+        message: "success",
+      });
     });
   }, {
-    query: z.object({ id: z.number() }),
-    response: { 204: z.object() },
+    params: Params,
+    response: {
+      204: Message,
+      404: Error,
+    },
   })
